task-manager-back/pyproject.toml
[project]
name = "task-manager-back"
version = "0.1.0"
description = ""
authors = [
    {name = "kessa99",email = "kipre325@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.0,<2.0.0)",
    "email-validator (>=2.0.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "task_manager_back", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
